+++ winaccent/__init__.py
@@ -8,13 +8,15 @@ __author__ = "Valer100"
 import sys, time
 
 if not sys.platform == "win32": 
-    raise ImportError("No Windows environment found. This module only works on Windows 8 and later.")
+    raise ImportError("No Windows environment found. This module only works on Windows Vista and later.")
 elif sys.getwindowsversion().major == 10: 
     from . import _win_10 as win
 elif sys.getwindowsversion().major == 6 and sys.getwindowsversion().minor >= 2:
     from . import _win_8 as win
+elif sys.getwindowsversion().major == 6 and sys.getwindowsversion().minor >= 0:
+    from . import _win_vista as win
 else:
-    raise ImportError("Incompatible Windows version. This module only works on Windows 8 and later.")
+    raise ImportError("Incompatible Windows version. This module only works on Windows Vista and later.")
 
 
 # Accent colors